+++ src/main/java/com/er/sagar/BlogAPI/DTO/PostLikesDto.java
@@ -7,10 +7,10 @@ import com.er.sagar.BlogAPI.entities.PostLike;
 @Component
 public class PostLikesDto {
 
-    private int id;
-    private int userId;
+    private Long id;
+    private Long userId;
     private String likedBy;
-    private int postId;
+    private Long postId;
 
     public PostLikesDto() {
 
@@ -23,19 +23,19 @@ public class PostLikesDto {
         this.userId = postLike.getUser().getId();
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
@@ -47,11 +47,11 @@ public class PostLikesDto {
         this.likedBy = likedBy;
     }
 
-    public int getPostId() {
+    public Long getPostId() {
         return postId;
     }
 
-    public void setPostId(int postId) {
+    public void setPostId(Long postId) {
         this.postId = postId;
     }
